requirements.txt
aiosqlite==0.21.0
fastapi==0.118.0
gradio==5.47.2
httpx==0.28.1
langchain==0.3.27
langchain_community==0.3.30
langchain_core==0.3.77
langchain-openai==0.3.34
pydantic==2.11.9
python-dotenv==1.1.1
SQLAlchemy==2.0.43
uvicorn==0.37.0